employee endpoint and more

+++ Solomonlol.ShiftLogger/Services/ShiftService.cs
@@ -1,5 +1,7 @@
-﻿using Microsoft.EntityFrameworkCore;
+﻿using AutoMapper;
+using Microsoft.EntityFrameworkCore;
 using ShiftLogger.Backend.Entities;
+using ShiftLogger.Backend.Entities.Dto;
 using ShiftLogger.Backend.Interfaces;
 using Solomonlol.ShiftLogger;
 
@@ -9,12 +11,14 @@ namespace ShiftLogger.Backend.Services
     public class ShiftService : IShiftService
     {
         private readonly ApplicationContext _context;
-        public ShiftService(ApplicationContext context)
+        private readonly IMapper _mapper;
+        public ShiftService(ApplicationContext context, IMapper mapper)
         {
             _context = context;
+            _mapper = mapper;
         }
 
-        public async Task Create(Shift shift, CancellationToken cancellationToken = default)
+        private async Task Create(Shift shift, CancellationToken cancellationToken = default)
         {
             await _context.Shifts.AddAsync(shift, cancellationToken);
             await _context.SaveChangesAsync(cancellationToken);
@@ -23,8 +27,34 @@ namespace ShiftLogger.Backend.Services
         public async Task Delete(int id, CancellationToken cancellationToken = default)
         {
             var shift = await _context.Shifts.FindAsync(id, cancellationToken);
-            _context.Shifts.Remove(shift);
+            if(shift!=null)
+                _context.Shifts.Remove(shift);
             await _context.SaveChangesAsync(cancellationToken);
+        }
+
+        public async Task End(int employeeNumber, ShiftDto dto, CancellationToken cancellationToken = default)
+        {
+            var employee = await _context.Employees.FirstOrDefaultAsync(e => e.EmployeeNumber == employeeNumber, cancellationToken);
+            if (employee != null)
+            {
+                var checkShift = await _context.Shifts.FirstOrDefaultAsync(s => s.IsEnded == false && s.EmployeeId == employee.Id, cancellationToken);
+                if (checkShift != null)
+                {
+                    await Update(_mapper.Map<Shift>(dto), cancellationToken);
+                }
+            }
+        }
+        public async Task Start(int employeeNumber, ShiftDto dto, CancellationToken cancellationToken = default)
+        {
+            var employee = await _context.Employees.FirstOrDefaultAsync(e => e.EmployeeNumber == employeeNumber, cancellationToken);
+            if (employee != null)
+            {
+                var checkShift = await _context.Shifts.FirstOrDefaultAsync(s => s.IsEnded == false && s.EmployeeId == employee.Id, cancellationToken);
+                if (checkShift == null)
+                {
+                    await Create(_mapper.Map<Shift>(dto), cancellationToken);
+                }
+            }
         }
 
         public async Task<IEnumerable<Shift>> GetAll(CancellationToken cancellationToken = default)
@@ -41,6 +71,7 @@ namespace ShiftLogger.Backend.Services
         {
             return await _context.Shifts.FindAsync(id, cancellationToken);
         }
+
 
         public async Task Update(Shift shift, CancellationToken cancellationToken = default)
         {
